Fix all example clippy warnings: 32 → 4 (only cfg(v4l2) platform-gates remain)
Fixed across 13 example files:
- Vec::new()+push → vec![] macro (3 files)
- from_str → from_name for DemosaicAlgo
- format! → .to_string() (2 files)
- unnecessary to_string() removed
- clone() → from_ref() (5 files)
- empty println!("") → println!()
- unused imports removed (std::ptr, std::io::Write)
- clamp-like patterns → .clamp()
- identical if blocks → plain value
- unwrap after is_ok → match/Ok()
- for loop indexing → .iter().enumerate()
- Default::default() field init → struct literal with ..Default::default()

Remaining 4 warnings: all cfg(v4l2) — platform-specific, expected.

Lib: 0 warnings. Tests: 540 pass. Doc: 0 warnings.

+++ cam-rust/cam-isp/examples/bench_8k_to_fhd.rs
@@ -62,7 +62,7 @@ fn main() {
     let full_h   = sensor_h as i64;
     let ds_w     = pipe_w as i64;
     let ds_h     = pipe_h as i64;
-    let mid_h    = (full_h / 2) as i64;   // 2160 (after UnpackCfa stride=2)
+    let mid_h    = full_h / 2;   // 2160 (after UnpackCfa stride=2)
     let mid_w    = packed_w;               // 3840
 
     let mut blocks: Vec<Box<dyn IspBlock>> = vec![
